training/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.41.0
diffusers>=0.29.0
accelerate>=0.30.0
datasets>=2.19.0
safetensors>=0.4.3
Pillow>=10.3.0
opencv-python-headless>=4.10.0.84
kaggle>=1.6.17
tqdm>=4.66.4
numpy>=1.26.4
scipy>=1.13.1
peft>=0.11.1
trl>=0.8.6
spacy>=3.7.4
sentencepiece>=0.2.0
xformers>=0.0.27.post2